get rid of glm

+++ src/camera.cpp
@@ -1,5 +1,5 @@
 #include "camera.h"
 
-ray camera::cast_ray(float u, float v) {
+ray camera::cast_ray(double u, double v) {
     return ray{origin, lower_left_corner + u * horizontal + v * vertical - origin};
 }
