+++ src/hittable.h
@@ -4,5 +4,5 @@ struct hit_record;
 struct ray;
 
 struct hittable {
-    virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
+    virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
 };
