+++ src/hittable_list.h
@@ -11,5 +11,5 @@ struct hittable_list: public hittable {
     void add(std::shared_ptr<hittable> object) { objects.push_back(object); }
     void clear() { objects.clear(); }
 
-    bool hit(const ray& r, float tmin, float tmax, hit_record& rec) const override;
+    bool hit(const ray& r, double tmin, double tmax, hit_record& rec) const override;
 };
