+++ src/sphere.h
@@ -10,11 +10,11 @@ struct material;
 struct sphere: public hittable {
     std::shared_ptr<material> mat_ptr;
     vec3 center;
-    float radius;
+    double radius;
 
-    sphere(const vec3& center, const float radius, std::shared_ptr<material> m) : center(center), radius(radius), mat_ptr(m) {}
+    sphere(const vec3& center, const double radius, std::shared_ptr<material> m) : center(center), radius(radius), mat_ptr(m) {}
 
-    bool hit(const ray& r, float tmin, float tmax, hit_record& rec) const override;
+    bool hit(const ray& r, double tmin, double tmax, hit_record& rec) const override;
 };
